Changes done in Entities and Repo

+++ src/main/java/com/ty/Shopit/Repo/CustomerOrderRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ty.Shopit.Entity.CustomerOrder;
 
-public interface CustomerOrderRepo extends JpaRepository<CustomerOrder, Integer>{
+public interface CustomerOrderRepo extends JpaRepository<CustomerOrder, Long>{
 
 }
