+++ src/main/java/com/ty/Shopit/Repo/ReviewRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ty.Shopit.Entity.Review;
 
-public interface ReviewRepo extends JpaRepository<Review, Integer>{
+public interface ReviewRepo extends JpaRepository<Review, Long>{
 
 }
